Move top level files into new `frontend` directory
