+++ src/frontend/anti_piracy_screen_mode.cpp
