+++ src/frontend/game_session_mode.cpp
